fix: resolve Float/Double type compatibility in rating components
- Remove unnecessary .toDouble() conversions in CompactStarRating calls
- Fix type mismatches where Double was passed but Float? expected
- Ensure proper type compatibility between rating components
- All compilation errors resolved, build now succeeds

+++ feature/playlist/src/main/java/com/deadarchive/feature/playlist/components/ReviewDetailsSheet.kt
@@ -196,7 +196,7 @@ private fun RatingSummaryCard(
             )
             
             CompactStarRating(
-                rating = rating.toDouble(),
+                rating = rating,
                 confidence = null,
                 starSize = IconResources.Size.MEDIUM
             )
@@ -293,7 +293,7 @@ private fun ReviewItem(
                     horizontalArrangement = Arrangement.spacedBy(4.dp)
                 ) {
                     CompactStarRating(
-                        rating = review.stars,
+                        rating = review.stars.toFloat(),
                         confidence = null,
                         starSize = IconResources.Size.SMALL
                     )
